🔥 category 파일 삭제
