--- src/components/common/category.js
@@ -1,7 +1,0 @@
-export const category = [
-  { key: "COUPLE", label: "커플스냅" },
-  { key: "FRIENDSHIP", label: "우정스냅" },
-  { key: "GRADUATION", label: "졸업스냅" },
-  { key: "WEDDING", label: "웨딩스냅" },
-  { key: "FAMILY", label: "가족스냅" },
-];
